feat: added basic bun server to serve static html
