+++ pages/creator.html
@@ -7,8 +7,8 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&family=Besley:wght@400;500;600&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="css/styles.css">
-    <link rel="stylesheet" href="css/creator.css">
+    <link rel="stylesheet" href="../css/styles.css">
+    <link rel="stylesheet" href="../css/creator.css">
 </head>
 <body>
     <div class="creator-container">
@@ -37,7 +37,7 @@
             </div>
             
             <div class="nav-link">
-                <a href="index.html">← Back to Gallery</a>
+                <a href="/">← Back to Gallery</a>
             </div>
         </aside>
         
@@ -49,9 +49,7 @@
         </main>
     </div>
     
-    <script src="js/storage.js"></script>
-    <script src="js/planet-generator.js"></script>
-    <script src="js/controls.js"></script>
+    <script type="module" src="../src/controls.ts"></script>
 </body>
 </html>
