+++ pages/index.html
@@ -7,8 +7,8 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&family=Besley:wght@400;500;600&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="css/styles.css">
-    <link rel="stylesheet" href="css/gallery.css">
+    <link rel="stylesheet" href="../css/styles.css">
+    <link rel="stylesheet" href="../css/gallery.css">
 </head>
 <body>
     <div class="gallery-container">
@@ -21,12 +21,11 @@
         </div>
         
         <div class="create-new-container">
-            <a href="creator.html" class="create-new-btn">+ Create New Planet</a>
+            <a href="/creator" class="create-new-btn">+ Create New Planet</a>
         </div>
     </div>
     
-    <script src="js/storage.js"></script>
-    <script src="js/gallery.js"></script>
+    <script type="module" src="../src/gallery.ts"></script>
 </body>
 </html>
